* Save file with websocket and fs in server file * Condense api.js with server.js * Make textarea and publish to save text to local filesystem
